gitignore bin and lib directories rather than .exe and .dll files
.exe and .dll are Windows-only.

Also, the convention seems to be that the Windows application manifest should be
named moac.exe.xml, which would have been ignored by the moac.* rule. Searching
the code base for "moac." finds "moac.dat" (already ignored by the *.dat rule)
and "moac.log". Added a *.log rule to cover this.

Removed one *.o rule as it was duplicated.