Add AppImage builds for portable linux install
This adds an AppImage builds. An AppImage is essentially a way to bundle
the game with all of our libraries into a folder that behaves more or
less like an executable with an entrypoint. In the linux world it's by
far the best performing (raw performance) way of bundling the game in a
single portable "file" that will run on any distro. In fact, it's also a
way that we can bundle in more highly optimized versions of the
libraries we use, than the ones that are typically installed on most
distros. In this way it can often outperform typical native installs.

This app image uses CachyOS which is a flavor of arch linux, which
defaults to -O3 optimized builds of all libraries automatically. This
should perform better over the default of -O2 that most distro mirrors
build their packages with.

Separately, this provides a helper make command for users with zen4 or
zen5 chips. It does not add this to the pipeline however, because doing
so would likely increase pipeline build time for little added benefit.

Having said that, both @eddoww and I have zen4/5 chips, so this is a
selfish local helper so we can have extra speedy local builds to play
with. This could be easily extended for other targets - for instance,
this could be quite easily extended for either x86_64-v3 or x86_64-v4
optimized builds.

If a developer wanted to do that, it would be a matter of simply taking
in a new flag for those, and using it to conditionally append those
mirrors in the Dockerfile for app image builds, and then pass their
desired --march flag on through for the astonia compilation itself.