Add apple app bundle and dmg build + signing
This adds new Makefile commands for locally building a proper .app bundle for iOS, with recursive dependencies added appropriately (skipping system libraries). It also adds support for locally signing and notarizing the app bundle, as well as automatically converting the .ico file to .icns, and signing and notarizing both the app bundle and dmg in CI release builds.